refact: Refactors Answer Page for modularity and state management
Introduces AnswerEditorContext to centralize state handling for the editor and refactors AnswerEditor to leverage this context. Adds modular components like SubmitButton, EditorWrite, and EditorPreview for improved code organization and usability.

Enhances Answer and Staking state management by introducing dedicated providers and hooks. Updates related components such as Question and Answers to use the new structure. Adds type definitions for improved type safety.

Improves maintainability by cleaning up unused code, restructuring files, and adopting a more modular architecture.
