+++ frontend/react/src/pages/AnswerPage/Answers/index.tsx
@@ -5,9 +5,9 @@ import React, { useContext, useState, Suspense } from "react"
 import { useAccount } from "@starknet-react/core"
 import { shortenAddress } from "@utils/shortenAddress"
 
-import { AnswersContext } from "../providers/AnswersProvider/answersContext"
+import { AnswersContext } from "../hooks/useAnswers/answersContext"
 
-import type { Question } from "../types"
+import type { Question } from "@app-types/index"
 import { useWallet } from "@hooks/useWallet"
 import { useStatusMessage } from "@hooks/useStatusMessage"
 
@@ -125,7 +125,7 @@ export function Answers({ question, setQuestion }: AnswersProps) {
           <p>No answers yet. Be the first to answer!</p>
         ) : (
           sortedAnswers.map((answer) => (
-            <AnswerItem key={answer.id} isCorrect={answer.isCorrect}>
+            <AnswerItem key={answer.id} $isCorrect={answer.isCorrect}>
               <AnswerHeader>
                 <UserAvatar
                   src={`https://avatars.dicebear.com/api/identicon/${answer.authorAddress}.svg`}
@@ -145,7 +145,7 @@ export function Answers({ question, setQuestion }: AnswersProps) {
               </AnswerHeader>
 
               <AnswerContent>
-              <Suspense fallback={<p>Carregando visualização...</p>}>
+                <Suspense fallback={<p>Carregando visualização...</p>}>
                   <ReactMarkdown
                     remarkPlugins={[remarkGfm]}
                     components={{
