+++ frontend/react/src/pages/AnswerPage/hooks/useStaking/staking.context.ts
