+++ frontend/react/src/types/answer.ts
@@ -1,0 +1,9 @@
+export type Answer = {
+  id: string
+  authorAddress: string
+  authorName: string
+  content: string
+  timestamp: string
+  isCorrect: boolean
+  votes: number
+}
