+++ frontend/react/src/types/contract.ts
@@ -1,0 +1,5 @@
+export interface ContractState {
+  isLoading: boolean
+  error: string | null
+  transactionHash: string | null
+}
