+++ frontend/react/src/types/index.ts
@@ -1,0 +1,4 @@
+export * from "./question"
+export * from "./answer"
+export * from "./abi"
+export * from "./contract"
